imu: change IMU to ICM-20689 (SPI)
Data readout works, but the code needs polishing. Especially imu_read and
imu_get_data procedures.
IMU works on SPI1 with a low speed clock of 500 kHz. This should be increased
as it allows for 3.3 kHz poll frequency only. We should aim towards 4 or 8 kHz.

+++ inc/imu.h
@@ -27,7 +27,7 @@ typedef struct {
   int16_t gyro_x;
   int16_t gyro_y;
   int16_t gyro_z;
-  int16_t temperature;
+  float temperature;
 } imu_data_t;
 
 extern imu_data_t imu_data;
